--- src/App/Contact/Contact.js
@@ -1,15 +1,0 @@
-import './Contact.css';
-
-function Contact() {
-
-    return (
-        <p className='contact-info'>
-            <span className='contact'> Contact: </span><br />
-            Megan Crotteau <br />
-            Denver, CO <br />
-            megan.crotteau@gmail.com <br />
-        </p>
-    )
-}
-
-export default Contact;
