+++ src/Contact/Contact.js
@@ -1,0 +1,17 @@
+import './Contact.css';
+
+function Contact() {
+
+    return (
+        <section className='contact-section'>
+            <p className='contact-info'>
+                <span className='contact'> Contact: </span><br />
+                Megan Crotteau <br />
+                Denver, CO <br />
+                megan.crotteau@gmail.com <br />
+            </p>
+        </section>
+    )
+}
+
+export default Contact;
